requirements.txt
django
djangorestframework
python-dotenv
# markdown       # Markdown support for the browsable API.
# django-filter  # Filtering support
stripe